requirements.txt
streamlit~=1.45.1
requests~=2.32.3
validators~=0.35.0
pillow~=11.2.1